update Kraken API version to include my 2 merged PRs
1 - volume is read from JSON value instead of parsed from description string (https://github.com/beldur/kraken-go-api-client/pull/34)
2 - sdk now uses float64 values for improved accuracy when parsing numbers (https://github.com/beldur/kraken-go-api-client/pull/35)

+++ support/kraken/getAccountBalances.go
@@ -28,10 +28,8 @@ func (k krakenExchange) GetAccountBalances(assetList []assets.Asset) (map[assets
 	return m, nil
 }
 
-// this currently returns a float32 and is not very accurate.
-// Waiting on my PR to change this to use float64: https://github.com/beldur/kraken-go-api-client/pull/35
-func getFieldValue(object krakenapi.BalanceResponse, fieldName string) float32 {
+func getFieldValue(object krakenapi.BalanceResponse, fieldName string) float64 {
 	r := reflect.ValueOf(object)
 	f := reflect.Indirect(r).FieldByName(fieldName)
-	return f.Interface().(float32)
+	return f.Interface().(float64)
 }
